refactor : onboarding api 성능 개선 #152 (#153)
* refactor : @Transactional 어노테이션 @AppService 계층으로 이동 및 register 메서드에서 onboarding 연관관계 추가

* test : BatchOnboardingCategoryRepository 테스트 - 실제 사용하지 x

* deprecated : 리팩토링 중 테스트 사용 - Deprecated 예정

* feat : 각 카테고리의 첫번째 아이템 쿼리 결과

* refactor : 각 카테고리의 첫번째 아이템 값 쿼리 메서드 명 변경 및 반환값 변경에 따른 리팩토링

* refactor : 각 카테고리의 첫번째 아이템 값 쿼리 메서드 변경 및 반환값 변경에 따른 리팩토링

* refactor :  각 카테고리의 첫번째 아이템 값 쿼리 메서드 변경 및 반환값 변경 - findFirstItemsByCategoryTypes

* test : findFirstByCategoryType 테스트 변경

* refactor : LEFT JOIN -> JOIN FETCH 로 N+1 문제 해결

* test : findFirstByCategoryType 테스트 수정

* refactor : findFirstByCategoryType 메서드 반환 타입 변경에 대한 리팩토링

* refactor : categroyType getter 메서드 추가

* refactor : NextWal -> NextWalEntity

* refactor : NextWalRepository 패키지 변경 및 리팩토링

* feat : saveAllInBatch 메서드 정의 및 구현

* feat : NextWal 영속성 관련 로직 분리 및 구현

* feat : findNextWalsByUserId 메서드 반환 DTO 클래스

* feat : NextWal 도메인 객체

* refactor : NextWalEntity 영속성 객체 <-> NextWal 도메인 객체 분리로 인한 변경

* refactor : 기존 NextWalService 제거 후 리팩토링

* feat : Item 영속성 연결 포트

* refactor : ItemRepository 패키지 변경 및 리팩토링

* refactor : CategoryRepository 패키지 변경 및 리팩토링

* feat : ItemPersistencePort 구현

* refactor : Item -> ItemEntity

* refactor : Category -> CategoryEntity

* refactor : ItemEntity 영속성 객체 <-> Item 도메인 객체 분리로 인한 변경

* refactor : CategoryEntity 영속성 객체 <-> Category 도메인 객체 분리로 인한 변경

* refactor : ItemRepository import -> ItemPersistencePort

* refactor : categoryItemNumber 타입 double -> int

* refactor : NextWalRepository 패키지 변경에 의한 변동

* config : enable batch insert, update

* refactor : COUNT_OF_CATEGORY_TYPE 타입 int 로 변경

* refactor : 상수 값 수정

* refactor : TestItemInitializer 리팩토링

* refactor : NextWalRepositoryTest 패키지 변경 및 리팩토링

* test : NextWalRepositoryCustomImplTest 테스트

* refactor : ItemRepositoryTest 패키지 변경 및 리팩토링

* refactor : category, item 엔티티 관련 패키지 변경에 의한 변동

* refactor : RandomRangeGenerator, RandomRangeGeneratorImpl 패키지 이동

* refactor : 정적 팩토리 메서드 명 변경 - from -> of

* refactor : GetNextWalUseCase, GetNextWalService -> Retrieve 로 변경

* feat : findNextWalsByUserId 쿼리 결과 매핑 dto

* refactor : 쿼리 결과 매핑 DTO 사용 시, 엔티티를 필드로 매핑할 경우 N+1 문제 발생 - 필요한 속성값만 전달하도록 변경

* refactor : findNextWalsByUserId 메서드 반환 값 변경에 따른 변경

* add : create 팩토리 메서드 추가

* add : create 팩토리 메서드 추가

+++ src/main/java/backend/wal/wal/item/application/port/in/RegisterItemRequestDto.java
@@ -4,9 +4,9 @@ public class RegisterItemRequestDto {
 
     private final String contents;
     private final String imageUrl;
-    private final double categoryItemNumber;
+    private final int categoryItemNumber;
 
-    public RegisterItemRequestDto(String contents, String imageUrl, double categoryItemNumber) {
+    public RegisterItemRequestDto(String contents, String imageUrl, int categoryItemNumber) {
         this.contents = contents;
         this.imageUrl = imageUrl;
         this.categoryItemNumber = categoryItemNumber;
@@ -20,7 +20,7 @@ public class RegisterItemRequestDto {
         return imageUrl;
     }
 
-    public double getCategoryItemNumber() {
+    public int getCategoryItemNumber() {
         return categoryItemNumber;
     }
 }
